+++ src/anime/anime_component.rs
