+++ src/anime/mod.rs
@@ -1,0 +1,6 @@
+pub mod anime_component;
+pub mod anime_res;
+pub mod attach_animations;
+pub mod insert_animations;
+pub mod overlay_animations;
+pub mod tile_animations;
