--- src/components/mod.rs
@@ -1,4 +1,0 @@
-pub mod animation;
-pub mod behavior_patterns;
-pub mod kinetic_entity;
-pub mod sound;
