+++ src/map/tiled.rs
@@ -14,8 +14,8 @@ use bevy_ecs_tilemap::map::TilemapTexture;
 use futures_lite::AsyncReadExt;
 use tiled::{DefaultResourceCache, Loader, ResourceReader};
 
+use crate::anime::anime_component::AnimationComponent;
 use crate::ASSETS_BASE_PATH;
-use crate::components::animation::AnimationComponent;
 
 #[derive(AssetCollection, Resource)]
 pub struct TiledMapAssets {
