--- src/resources/mod.rs
@@ -1,3 +1,0 @@
-pub mod animation;
-pub mod audio;
-pub mod tiled;
