--- src/systems/animation_loadtime/mod.rs
@@ -1,2 +1,0 @@
-pub mod attach_animations;
-pub mod insert_animations;
