--- src/systems/animation_runtime/mod.rs
@@ -1,2 +1,0 @@
-pub mod overlay_animations;
-pub mod tile_animations;
