--- src/systems/entity_loadtime/mod.rs
@@ -1,4 +1,0 @@
-pub mod camera;
-pub mod environment_entity;
-pub mod player_entity;
-pub mod tiled_map;
