--- src/systems/entity_runtime/mod.rs
@@ -1,3 +1,0 @@
-pub mod camera;
-pub mod environment_entity;
-pub mod player_entity;
