--- src/systems/mod.rs
@@ -1,5 +1,0 @@
-pub mod animation_loadtime;
-pub mod animation_runtime;
-pub mod audio_runtime;
-pub mod entity_loadtime;
-pub mod entity_runtime;
